refactor: move internal wire types to reduce circular dependencies
