+++ packages/transactions/src/address.ts
@@ -1,36 +1,5 @@
-import {
-  AddressHashMode,
-  AddressVersion,
-  RECOVERABLE_ECDSA_SIG_LENGTH_BYTES,
-  StacksWireType,
-} from './constants';
-
-import { c32address } from 'c32check';
-import { hexToBytes } from '@stacks/common';
 import { TransactionVersion } from '@stacks/network';
-
-export interface AddressWire {
-  readonly type: StacksWireType.Address;
-  readonly version: AddressVersion;
-  readonly hash160: string;
-}
-
-export interface MessageSignatureWire {
-  readonly type: StacksWireType.MessageSignature;
-  data: string;
-}
-
-export function createMessageSignature(signature: string): MessageSignatureWire {
-  const length = hexToBytes(signature).byteLength;
-  if (length != RECOVERABLE_ECDSA_SIG_LENGTH_BYTES) {
-    throw Error('Invalid signature');
-  }
-
-  return {
-    type: StacksWireType.MessageSignature,
-    data: signature,
-  };
-}
+import { AddressHashMode, AddressVersion } from './constants';
 
 /**
  * Translates the tx auth hash mode to the corresponding address version.
@@ -71,12 +40,4 @@ export function addressHashModeToVersion(
     default:
       throw new Error(`Unexpected hashMode ${JSON.stringify(hashMode)}`);
   }
-}
-
-export function addressFromVersionHash(version: AddressVersion, hash: string): AddressWire {
-  return { type: StacksWireType.Address, version, hash160: hash };
-}
-
-export function addressToString(address: AddressWire): string {
-  return c32address(address.version, address.hash160);
 }
