+++ packages/transactions/tests/builder.test.ts
@@ -19,14 +19,20 @@ import {
   CONTRACT_ABI_PATH,
   ClarityAbi,
   READONLY_FUNCTION_CALL_PATH,
+  StacksWireType,
   TRANSACTION_FEE_ESTIMATE_PATH,
+  TokenTransferPayloadWire,
   TxBroadcastResult,
   TxBroadcastResultOk,
   TxBroadcastResultRejected,
   addressFromPublicKeys,
+  addressToString,
   broadcastTransaction,
   callReadOnlyFunction,
+  createMessageSignature,
   createStacksPublicKey,
+  createTokenTransferPayload,
+  createTransactionAuthField,
   estimateFee,
   estimateTransaction,
   getContractMapEntry,
@@ -35,9 +41,11 @@ import {
   privateKeyToPublic,
   publicKeyIsCompressed,
   publicKeyToHex,
+  serializePayloadBytes,
   serializePostConditionBytes,
   serializePublicKeyBytes,
 } from '../src';
+import { BytesReader } from '../src/BytesReader';
 import {
   MultiSigSpendingCondition,
   SingleSigSpendingCondition,
@@ -59,7 +67,6 @@ import {
   makeUnsignedSTXTokenTransfer,
   sponsorTransaction,
 } from '../src/builders';
-import { BytesReader } from '../src/bytesReader';
 import {
   ClarityType,
   UIntCV,
@@ -72,7 +79,6 @@ import {
   standardPrincipalCV,
   uintCV,
 } from '../src/clarity';
-import { addressToString, createMessageSignature } from '../src/common';
 import {
   AddressHashMode,
   AuthType,
@@ -80,17 +86,10 @@ import {
   PayloadType,
   PostConditionMode,
   PubKeyEncoding,
-  StacksWireType,
   TxRejectedReason,
 } from '../src/constants';
 import { makeRandomPrivKey } from '../src/keys';
-import {
-  TokenTransferPayloadWire,
-  createTokenTransferPayload,
-  serializePayloadBytes,
-} from '../src/payload';
 import { FungiblePostCondition, PostCondition, StxPostCondition } from '../src/postcondition-types';
-import { createTransactionAuthField } from '../src/signature';
 import { TransactionSigner } from '../src/signer';
 import {
   StacksTransaction,
